wip: big refactor / per-listener change of translator
