+++ internal/control-plane/eventloop/eventloop.go
@@ -218,55 +218,3 @@ func (e *eventLoop) updateXds(snap *snapshot.Cache) {
 	}
 }
 
-//// gets the subset of upstreams which are destinations for at least one route in at least one
-//// virtual service
-//func destinationUpstreams(allUpstreams []*v1.Upstream, virtualServices []*v1.VirtualService) []*v1.Upstream {
-//	destinationUpstreamNames := make(map[string]bool)
-//	for _, vs := range virtualServices {
-//		for _, route := range vs.Routes {
-//			dests := getAllDestinations(route)
-//			for _, dest := range dests {
-//				var upstreamName string
-//				switch typedDest := dest.DestinationType.(type) {
-//				case *v1.Destination_Upstream:
-//					upstreamName = typedDest.Upstream.Name
-//				case *v1.Destination_Function:
-//					upstreamName = typedDest.Function.UpstreamName
-//				default:
-//					panic("unknown destination type")
-//				}
-//				destinationUpstreamNames[upstreamName] = true
-//			}
-//		}
-//	}
-//	var destinationUpstreams []*v1.Upstream
-//	for _, us := range allUpstreams {
-//		if _, ok := destinationUpstreamNames[us.Name]; ok {
-//			destinationUpstreams = append(destinationUpstreams, us)
-//		}
-//	}
-//	return destinationUpstreams
-//}
-//
-//func getAllDestinations(route *v1.Route) []*v1.Destination {
-//	var dests []*v1.Destination
-//	if route.SingleDestination != nil {
-//		dests = append(dests, route.SingleDestination)
-//	}
-//	for _, dest := range route.MultipleDestinations {
-//		dests = append(dests, dest.Destination)
-//	}
-//	return dests
-//}
-//
-//func destinationEndpoints(upstreams []*v1.Upstream, allEndpoints endpointdiscovery.EndpointGroups) endpointdiscovery.EndpointGroups {
-//	destinationEndpoints := make(endpointdiscovery.EndpointGroups)
-//	for _, us := range upstreams {
-//		eps, ok := allEndpoints[us.Name]
-//		if !ok {
-//			continue
-//		}
-//		destinationEndpoints[us.Name] = eps
-//	}
-//	return destinationEndpoints
-//}
